Added delete verification with sweetalert

+++ src/components/TaskList.jsx
@@ -1,17 +1,30 @@
 import axios from 'axios';
 import Ticker from './Ticker';
+import swal from 'sweetalert';
 
 function TaskList ({key, task, fetch}) {
  
 const display = task.complete ? 'completed' : 'notCompleted';
 
 const deleteTask = () => {
-axios.delete(`todo/${task.id}`)
-.then((response) => {
-    fetch();
-}).catch((error) => {
-    console.log(error);
-})
+    swal({
+        title: `Delete ${task.title}?`,
+        text: 'Are you sure?',
+        icon: "warning",
+        buttons: true,
+        dangerMode: true
+        })
+        .then((value) => {
+            if (value) {
+            axios.delete(`todo/${task.id}`)
+            .then((response) => {
+                fetch();
+            }).catch((error) => {
+                console.log(error);
+            })
+        }
+    }
+        )
 }
 
     return (
